+++ src/libYARP_dev/src/DeviceDriver.cpp
@@ -68,7 +68,7 @@ bool DeviceResponder::read(ConnectionReader& connection) {
         ConnectionWriter *writer = connection.getWriter();
         if (writer!=nullptr) {
             if (response.get(0).toString()=="many"&&writer->isTextMode()) {
-                for (int i=1; i<response.size(); i++) {
+                for (size_t i=1; i<response.size(); i++) {
                     Value& v = response.get(i);
                     if (v.isList()) {
                         v.asList()->write(*writer);
